# api-gateway/cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hire_Jobs is a platform to find your dream job.",
        "title": "Hireo Jobs API",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/employer/chats": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves chat details based on the provided request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Retrieve chat details",
                "parameters": [
                    {
                        "description": "Chat request details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved chat details",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Details not in correct format\" or \"User ID not found in JWT claims\" or \"Failed to get chat details",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/group/:groupID/chat": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Processes WebSocket messages for group chat based on the provided group ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Process WebSocket group chat messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "groupID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "WebSocket connection established",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Missing Authorization header\" or \"Invalid token\" or \"Websocket Connection Issue\" or \"Error reading WebSocket message\" or \"Details not in correct format",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.ChatRequest": {
            "type": "object",
            "required": [
                "friendID",
                "limit",
                "offset"
            ],
            "properties": {
                "friendID": {
                    "type": "string"
                },
                "limit": {
                    "type": "string"
                },
                "offset": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerTokenAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}